ORKNavigableOrderedStack: use task results for backwards navigation
This commits removes ORKNavigableOrderedTask's '_stepIdentifierStack' ivar (used for backwards navigation) and uses the managed task results instead. The managed task result will only show results for the steps that have been previously visited, so it can be used to simulate the backwards navigation stack.

The ORKTaskViewController's '_managedStepIdentifiers' ivar has been converted into an 'NSMutableOrderedSet'. When an already visited step is shown (either by going back, or by following a loop in a navigable ordered task), all the managed step identifiers which come later than the step being shown are removed.

This commit also updates the 'ORKTest' to remove the "Save For Later" task archiving (which is no longer needed) and updates the 'Ordered Loop Task' example to be more comprehensive by featuring a branched question.