ORKTaskViewController: make _managedStepIdentifiers array again
This is the original behavior, it follows the pop/push model. All the backward navigation stack is preserved, but older results which are ovewritten if you revisit the same step on a loop. Thus, going over a loop will produce duplicate results for the steps that are seen more than once (all the duplicate step results will point to the same result instance: the one corresponding to the last time you visited the step).