Accessibility fix value picker
From Ryan:
This next issue is with VoiceOver speaking the wrong words when changing ORKValuePicker’s value.  Ultimately, we need to fix this on our side:
<rdar://problem/23885294> UITextField AX broken when connected as the firstResponder to a UIPicker.

I have attached a workaround patch to rdar://problem/23710698. The problem is that when we change the value of the picker and a text field is the firstResponder, VoiceOver thinks this is a selection changed event and reads the delta change. For example, going from Male to Female, the delta is “LE.” The workaround is, since this textfield is not editable, to keep the selectedTextRange as 0,0 so that VO doesn’t do the selection changed behavior.

Please let me know if you have any questions!