ORKFormItemTextFieldBasedCell: Allow temporarily leaving the text field empty without triggering answer validity checking
This fixes a regression. It restores the behavior in which you are allowed to edit a different text field in a form if you leave the current text field blank. E.g., on ORKTest's 'Mini Form', tapping on the first cell text field and then tapping on second cell text field without having written anything should not trigger the alert.