Work In Progress
Created more genericised, Swift-friendly convenience inits for
NSPredicate to support our ORKResultPredicates. These now appear to
work OK in Swift.

Started moving the original Objective C functions across as veneers to
call the Swift code. We have an issue with choice ORKResults as they
aren’t guaranteed to be using a String (issue raised on RK group for
suggestions as to best solution).

This commit therefore won’t work with Obj C applications as many of the
inits are missing.