Added a series of Predicate tests to ORKTest. This should test a range of different predicate conditions and enable us to ensure that Predicates are working as expected.
NB predicate tests which use 'match' don't seem to be working as expected in these tests. In the choice match test we try to match against Cat*; but only an exact match for Cat seems to pass the test.

Test 7 - to check a text string for a match *og causes a SigAbt. This may be that the match strings are incorrect for these tests, or it may be an issue with the match predicate functions.