Fixing bug where timer label did not display
in the updateTimerLabel function, the timer would be hidden if the labelString was null. The labelString was null everytime because the "formatter" object (an NSDateComponentsFormatter object) was null the whole time. This object was null because inside the "dispatch_once", we did not reference/edit the existing static formatter object outside of the dispatch_once, but rather, we instantiated a new formatter object.