Add a flag to allow displaying the continue button in the standard position for the compltion step.
We are finding that many of our users are confused as to how to dismiss a task because the continue button is moved to the upper right corner for the last step.  This commit allows overriding the default behavoir of moving that button and places it in the standard position for a continue button.