Fix audio step duplicate file creation bug
Start was getting called in viewDidAppear:, creating an audio file, and meanwhile the superclass viewDidAppear: had dispatched a block to main that then saw the recorder was started and called resume, creating another audio file, only one of which was included in the results, leaving the other which the caller didn't know about and wasn't aware needed to be deleted, using up storage space over time. The fix is to use the shouldStartTimerAutomatically flag to let the superclass viewDidAppear: start the recorder, rather than override that method in the audio step view controller to start it explicitly.