[ORKESerialization] Fix ORKTextAnswerFormat tests
Adds NSRegularExpression serialization/deserialization. Note that the tests are ignoring the 'validationRegularExpression' for the hash test becuase it's purposefully not included there.

Also, slight fix to the CLCircularRegion serialization (a nil CLCircularRegion must be serialized into an empty dict, not into a nil object). Otherwise it'll crash if you try to add nil to the parent serialization dictionary.