Add Go-no-go active task
Developed by Roland Rabien. Submitted by Sage Bionetworks in collaboration with BiAffect.

In a go-no-go task, the participant is asked to move the device sharply in any direction in response to a visual cue. You can use this task to accurately assess the participant's simple reaction time. Randomly they will be given a cue to not shake the device; this tests impulsiveness.

A go-no-go time task finishes when the participant has completed the required number of attempts successfully. For a "go", an attempt is successful when the participant exerts acceleration greater than `thresholdAcceleration` to the device after the stimulus has been delivered and before `timeout` has elapsed. An attempt is unsuccessful if acceleration greater than `thresholdAcceleration` is applied to the device before the stimulus or if this does not occur before `timeout` has elapsed. If unsuccessful, the result is not reported and the participant must try again to proceed with the task. For a "no go", the sucess condition is reversed.

Data collected by the task is in the form of `ORKGoNoGoResult` objects. The researcher can use these results to evaluate the response to the stimulus, and calculate the reaction time.